Moving native definitions to fast-tooling
