+++ packages/tooling/fast-tooling/src/definitions/native/html-native.definition.ts
@@ -1,12 +1,12 @@
-import { DataType } from "@microsoft/fast-tooling";
+import { DataType } from "../../data-utilities";
 import {
     WebComponentAttribute,
     WebComponentAttributeValues,
     WebComponentDefinition,
     WebComponentDefinitionTag,
     WebComponentSlot,
-} from "@microsoft/fast-tooling/dist/esm/data-utilities/web-component";
-import { voidElements } from "@microsoft/fast-tooling/dist/esm/data-utilities/html-element";
+} from "../../data-utilities/web-component";
+import { voidElements } from "../../data-utilities/html-element";
 /**
  * These native html tag definitions are dependent on a devDependency on "vscode-web-custom-data"
  * specifically the browsers.html-data.json file which defines native html elements and their attributes
@@ -73,6 +73,7 @@ function convertAttributeData(tag: VSCodeNativeHTMLTag): WebComponentAttribute[]
             default: "",
             required: false,
             values: findValueSetValues(attribute.valueSet),
+            description: attribute.description.toString(),
         };
     });
 }
