+++ packages/tooling/fast-tooling/src/definitions/native/html-native.vs-code-v1.1-types.ts
